commit for start screen and about scree
