+++ index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Lemon Clicker - CSE 110 Lab 3</title>
+    <title>CSE110 Project Team 24 - GEOQUEST</title>
     <style>
       * {
         margin: 0;
